Fix signals for deleting images for crew and train models

+++ station_api/signals.py
@@ -8,11 +8,13 @@ from station_api.models import Crew, Train
 
 @receiver(pre_delete, sender=Crew)
 def delete_crew_image(sender, instance, **kwargs) -> None:
-    if os.path.isfile(instance.crew_image.path):
-        os.remove(instance.crew_image.path)
+    if instance.crew_image and hasattr(instance.crew_image, "path"):
+        if os.path.isfile(instance.crew_image.path):
+            os.remove(instance.crew_image.path)
 
 
 @receiver(pre_delete, sender=Train)
 def delete_train_image(sender, instance, **kwargs) -> None:
-    if os.path.isfile(instance.train_image.path):
-        os.remove(instance.train_image.path)
+    if instance.train_image and hasattr(instance.train_image, "path"):
+        if os.path.isfile(instance.train_image.path):
+            os.remove(instance.train_image.path)
